refactor(clock): IClock API를 uint64_t 나노초 정수로 통일
- clock.h namespace 오타(psh;::lib) 수정
- SteadyClock: double 초 → uint64_t 나노초로 내부 저장 단위 변경,
  생성자 파라미터를 std::chrono::nanoseconds 로 타입 안전화
- SteadyTickClock/ManualClock: override 시그니처를 uint64_t 로 맞춤
- clock_test: SteadyClock 생성자에 chrono literal 주입

IClock 인터페이스가 uint64_t 반환인데 구현체가 double 로 override 해
silent truncation + override 실패가 섞여 있던 상태를 정리.

+++ server/lib/include/clock/manual_clock.h
@@ -8,40 +8,40 @@ namespace psh::lib
 class ManualClock final : public IClock
 {
   public:
-    double GetCurrentTime() const override
+    uint64_t GetCurrentTime() const override
     {
         return currentTime_;
     }
-    double GetDeltaTime() const override
+    uint64_t GetDeltaTime() const override
     {
         return deltaTime_;
     }
-    double GetIdleTime() const override
+    uint64_t GetIdleTime() const override
     {
         return idleTime_;
     }
 
-    void SetCurrentTime(double time)
+    void SetCurrentTime(uint64_t time)
     {
         deltaTime_ = time - currentTime_;
         currentTime_ = time;
     }
 
-    void Advance(double seconds)
+    void Advance(uint64_t seconds)
     {
         currentTime_ += seconds;
         deltaTime_ = seconds;
     }
 
-    void SetIdleTime(double idle)
+    void SetIdleTime(uint64_t idle)
     {
         idleTime_ = idle;
     }
 
   private:
-    double currentTime_ = 0.0;
-    double deltaTime_ = 0.0;
-    double idleTime_ = 0.0;
+    uint64_t currentTime_ = 0.0;
+    uint64_t deltaTime_ = 0.0;
+    uint64_t idleTime_ = 0.0;
 };
 
 } // namespace psh::lib
